requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cachetools==5.3.3
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
charset-normalizer==3.3.2
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
Craft-xai==0.0.3
cycler==0.11.0
docopt==0.6.2
filelock==3.12.2
flatbuffers==1.12
fonttools==4.38.0
fsspec==2023.1.0
ftfy==6.1.1
gast==0.3.3
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
huggingface-hub==0.16.4
idna==3.6
imageio==2.31.2
importlib-metadata==6.7.0
joblib==1.3.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
mkl-fft==1.0.12
mkl-random==1.1.1
networkx==2.6.3
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.3.0.38
opt-einsum==3.3.0
packaging==24.0
Pillow==9.3.0
pipreqs==0.4.13
protobuf==3.20.3
psutil==6.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyWavelets==1.3.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.2
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.15.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.12
torch==1.7.1
torch-summary==1.4.5
torchaudio==0.7.0a0+a853dff
torchvision==0.8.2
tqdm==4.66.2
typing-extensions==3.7.4.3
urllib3==2.0.7
wcwidth==0.2.13
Werkzeug==2.2.3
wrapt==1.12.1
yarg==0.1.10
zipp==3.15.0
